// src/api/search/albumResult.ts
export interface AlbumSearchResult {
  result: Result
  code: number
}

export interface Result {
  hlWords: string[]
  albums: Album[]
  albumCount: number
}

export interface Album {
  name: string
  id: number
  type: string
  size: number
  picId: number
  blurPicUrl: string
  companyId: number
  pic: number
  picUrl: string
  publishTime: number
  description: string
  tags: string
  company?: string
  briefDesc: string
  artist: Artist
  songs: unknown
  alias: string[]
  status: number
  copyrightId: number
  commentThreadId: string
  artists: Artist2[]
  paid: boolean
  onSale: boolean
  picId_str?: string
  alg: string
  mark: number
  containedSong: string
}

export interface Artist {
  name: string
  id: number
  picId: number
  img1v1Id: number
  briefDesc: string
  picUrl: string
  img1v1Url: string
  albumSize: number
  alias: string[]
  trans: string
  musicSize: number
  topicPerson: number
  picId_str?: string
  transNames?: string[]
  img1v1Id_str?: string
  alia: string[]
}

export interface Artist2 {
  name: string
  id: number
  picId: number
  img1v1Id: number
  briefDesc: string
  picUrl: string
  img1v1Url: string
  albumSize: number
  alias: unknown[]
  trans: string
  musicSize: number
  topicPerson: number
  img1v1Id_str?: string
}